add fuzzing harnesses
Toxsave harness ported to libFuzzer interface.
New harness for bootstrap phase.
